microservices_http/ratelimit_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
redis==5.0.8
grpcio
grpcio-tools
protobuf
